Add files via upload
Edward: I had to make this repository public in order to create a branch. While it was private it refused to make a new branch. We can leave it public. Now I am trying to load synergy.R with some changes into my new branch. It may end up in the master repository (I hope not) and may override your latest changes (I really hope not).
The purpose was to add some output commands, e.g. one that makes a variance-covariance matrix into the corresponding correlation matrix. I needed those outputs while I was writing the paper. I also did a bit of commenting.
If the file ends up in my branch and not the master I will then download the corresponding data_info.R file which contains the command Y_0 <- 0.046404, a line which may be missing in your latest.