Add files via upload
It worked. Eventually I will run the suite several times with different values of Y_0: 0.025, 0.035, 0.046404, and 0.055. Because the choice of what zero-dose controls to include is legitimately controversial I think we should check that none of our essential conclusions about synergy  is affected by such changes in Y_0, which cover the range of different modeling attempt usages for the HG Bevalac/NSRL data set. Of course the details all change whenever Y_0 changes, including NTE-also vs. TE-only models that some people think is crucial but to my mind is merely a detail and is in any case almost irrelevant to synergy modeling since it is concerned entirely with 1-ion models, not mixtures.