Add files via upload
the main changes are for WebSup plots, e.g. our QQ plot whose (short) chunk is temporarily located near the end of plots.R. I think the plan is to have a special file for WebSup graphs. However the WebSup graphs use most of the whole suite, and I am now reducing the number of WebSup fgures, so we should consider just adding the WebSup figures as a big  chunk at the end of plots.R, where there are two figures at the moment. Other code changes are minor and there are also commenting changes.