Add files via upload
This is the latest version of my sandbox plots.R. It still works fine as long as Y_0 is no bigger than 0.0464 in data_info. But I have not yet had time to remove obsolete commenting, add relevant commenting; or clean up the naming.  The Monte Carlo commands in plots.R are the real bottleneck as regards (optional for LSSR though perhaps needed later and in any case worrisome) explorations of Y_0 = 0.05.  So I thought it best for Edward to have it available while he thinks about the uniroot bug(s).